add huobi pkg

+++ fetchhuobi.py
@@ -17,18 +17,25 @@ def fetch_save(symbol, csv_file):
     # df = pd.DataFrame(data, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'], dtype=np.float)
     # df.to_csv(path_to_csv)
 
-    data = pd.read_csv(path_to_csv)
-    last_ts = data.iloc[-1]['timestamp']
+    if os.path.exists(path_to_csv):
+        data = pd.read_csv(path_to_csv)
+        last_ts = data.iloc[-1]['timestamp']
+    else:
+        last_ts = 1602896400000 - 3600 * 1000
+
     while True:
         time.sleep(1)
-        klines = exchange.fetchOHLCV(symbol, '1h', since=(last_ts + 3600 * 1000))
+        klines = exchange.fetchOHLCV(symbol, '1h', since=(last_ts + 3600 * 1000), limit=2000)
         if len(klines) == 0:
             break
         else:
             last_ts = klines[-1][0]
             df = pd.DataFrame(klines, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'], dtype=np.float)
             df = df.set_index('timestamp')
-            df.to_csv(path_to_csv, mode='a', header=False)
+            if os.path.exists(path_to_csv):
+                df.to_csv(path_to_csv, mode='a', header=False)
+            else:
+                df.to_csv(path_to_csv, mode='w', header=True)
 
 
 def clean_csv(csv_file):
@@ -43,3 +50,4 @@ def clean_csv(csv_file):
 
 fetch_save('BTC/USDT', 'Huobi_BTCUSDT_1h.csv')
 fetch_save('BTC3S/USDT', 'Huobi_BTC3S_1h.csv')
+fetch_save('BTC1S/USDT', 'Huobi_BTC1S_1h.csv')
